+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Made `_kind` enumerable again.
+
 ## [0.2.0] - 2024-02-28
 
 - Made the first-defined node the start node for the grammar. It no longer